+++ src/any_reader.rs
@@ -48,7 +48,7 @@ impl<R: Read> SyncAnyPcapReaderInner<R> {
             )
         })?;
         if peak.len() < 4 {
-            return Err(std::io::Error::new(
+            Err(std::io::Error::new(
                 std::io::ErrorKind::UnexpectedEof,
                 "Failed to read magic number",
             ))?;